RecordDataを共通化してimport先を修正
