+++ src/app/api/records/route.ts
@@ -1,6 +1,7 @@
 import { prisma } from "@/app/_libs/prisma";
 import { getProfileByUserId } from "@/app/_libs/prisma/profile";
 import { getAuthenticatedUser } from "@/app/_libs/supabase/auth";
+import { RecordData } from "@/types/record";
 import type { Record } from "@prisma/client";
 import { NextRequest, NextResponse } from "next/server";
 
@@ -15,21 +16,13 @@ export type CreateRecordResponse = {
   id: number;
 };
 
-export type RecordsData = {
-    id: number
-    date: string
-    weight: number | null
-    steps: number | null
-    memo: string | null
-};
-
 export type RecordsResponse = {
   status: "OK" | "NG"
   message: string
-  records: RecordsData[]
+  records: RecordData[]
 }
 
-const formatRecords = (records: Record[]): RecordsData[] => {
+const formatRecords = (records: Record[]): RecordData[] => {
   return records.map((record) => ({
     id: record.id,
     date: record.date.toISOString(),
