+++ src/types/record.ts
@@ -1,0 +1,7 @@
+export type RecordData = {
+  id: number
+  date: string
+  weight: number | null
+  steps: number | null
+  memo: string | null
+}
